refactor(#293): extract useLeaderboard hook from leaderboard page
Move all state, fetch logic, and event handlers to use-leaderboard.ts.
Extract LeaderboardRow as a standalone component.
page.tsx is now pure JSX (~200 lines, down from 405).
